+++ kubernetes/resource_kubernetes_config_map_test.go
@@ -235,7 +235,10 @@ func testAccCheckConfigMapData(m *api.ConfigMap, expected map[string]string) res
 }
 
 func testAccCheckKubernetesConfigMapDestroy(s *terraform.State) error {
-	conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+	conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "kubernetes_config_map" {
@@ -263,7 +266,11 @@ func testAccCheckKubernetesConfigMapExists(n string, obj *api.ConfigMap) resourc
 			return fmt.Errorf("Not found: %s", n)
 		}
 
-		conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+		conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+		if err != nil {
+			return err
+		}
+
 		namespace, name, err := idParts(rs.Primary.ID)
 		if err != nil {
 			return err