+++ kubernetes/resource_kubernetes_cron_job.go
@@ -43,7 +43,10 @@ func resourceKubernetesCronJob() *schema.Resource {
 }
 
 func resourceKubernetesCronJobCreate(d *schema.ResourceData, meta interface{}) error {
-	conn := meta.(*KubeClientsets).MainClientset
+	conn, err := meta.(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	metadata := expandMetadata(d.Get("metadata").([]interface{}))
 	spec, err := expandCronJobSpec(d.Get("spec").([]interface{}))
@@ -70,7 +73,10 @@ func resourceKubernetesCronJobCreate(d *schema.ResourceData, meta interface{}) e
 }
 
 func resourceKubernetesCronJobUpdate(d *schema.ResourceData, meta interface{}) error {
-	conn := meta.(*KubeClientsets).MainClientset
+	conn, err := meta.(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	namespace, _, err := idParts(d.Id())
 	if err != nil {
@@ -102,7 +108,10 @@ func resourceKubernetesCronJobUpdate(d *schema.ResourceData, meta interface{}) e
 }
 
 func resourceKubernetesCronJobRead(d *schema.ResourceData, meta interface{}) error {
-	conn := meta.(*KubeClientsets).MainClientset
+	conn, err := meta.(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	namespace, name, err := idParts(d.Id())
 	if err != nil {
@@ -158,7 +167,10 @@ func resourceKubernetesCronJobRead(d *schema.ResourceData, meta interface{}) err
 }
 
 func resourceKubernetesCronJobDelete(d *schema.ResourceData, meta interface{}) error {
-	conn := meta.(*KubeClientsets).MainClientset
+	conn, err := meta.(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	namespace, name, err := idParts(d.Id())
 	if err != nil {
@@ -194,7 +206,10 @@ func resourceKubernetesCronJobDelete(d *schema.ResourceData, meta interface{}) e
 }
 
 func resourceKubernetesCronJobExists(d *schema.ResourceData, meta interface{}) (bool, error) {
-	conn := meta.(*KubeClientsets).MainClientset
+	conn, err := meta.(KubeClientsets).MainClientset()
+	if err != nil {
+		return false, err
+	}
 
 	namespace, name, err := idParts(d.Id())
 	if err != nil {