+++ kubernetes/resource_kubernetes_endpoints_test.go
@@ -275,7 +275,10 @@ func testAccCheckEndpointSubsets(svc *api.Endpoints, expected []api.EndpointSubs
 }
 
 func testAccCheckKubernetesEndpointDestroy(s *terraform.State) error {
-	conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+	conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "kubernetes_endpoints" {
@@ -305,7 +308,10 @@ func testAccCheckKubernetesEndpointExists(n string, obj *api.Endpoints) resource
 			return fmt.Errorf("Not found: %s", n)
 		}
 
-		conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+		conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+		if err != nil {
+			return err
+		}
 
 		namespace, name, err := idParts(rs.Primary.ID)
 		if err != nil {