+++ kubernetes/resource_kubernetes_role_test.go
@@ -283,7 +283,10 @@ func testAccCheckKubernetesRoleExists(n string, obj *api.Role) resource.TestChec
 			return fmt.Errorf("Not found: %s", n)
 		}
 
-		conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+		conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+		if err != nil {
+			return err
+		}
 
 		namespace, name, err := idParts(rs.Primary.ID)
 		if err != nil {
@@ -301,7 +304,10 @@ func testAccCheckKubernetesRoleExists(n string, obj *api.Role) resource.TestChec
 }
 
 func testAccCheckKubernetesRoleDestroy(s *terraform.State) error {
-	conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+	conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "kubernetes_role" {