+++ kubernetes/resource_kubernetes_stateful_set_test.go
@@ -311,7 +311,10 @@ func TestAccKubernetesStatefulSet_update_pod_template(t *testing.T) {
 }
 
 func testAccCheckKubernetesStatefulSetDestroy(s *terraform.State) error {
-	conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+	conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+	if err != nil {
+		return err
+	}
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "kubernetes_stateful_set" {
@@ -341,7 +344,10 @@ func testAccCheckKubernetesStatefulSetExists(n string, obj *api.StatefulSet) res
 			return fmt.Errorf("Not found: %s", n)
 		}
 
-		conn := testAccProvider.Meta().(*KubeClientsets).MainClientset
+		conn, err := testAccProvider.Meta().(KubeClientsets).MainClientset()
+		if err != nil {
+			return err
+		}
 
 		namespace, name, err := idParts(rs.Primary.ID)
 		if err != nil {